Moved DisplayBlock and DisplayParagraph out of Display for clarity
